#1467: Pre-filter mapping tables for HyperConformance * Addressed PR comments * Mapping Rule interpreters moved to own subfolder * More common code unification * Changed Warning to Exception when filter from configuration is not parse-able * Addressed compilation warning
