+++ spark-jobs/src/test/resources/log4j.properties
@@ -24,8 +24,8 @@ log4j.appender.general.layout.ConversionPattern=@log.pattern@
 # Suppress warnings logged within the code inside the tests
 log4j.logger.za.co.absa.enceladus.standardization.interpreter.stages.TypeParser$=@log.level.testcode@
 log4j.logger.za.co.absa.enceladus.conformance.interpreter.OptimizerTimeTracker=@log.level.testcode@
-log4j.logger.za.co.absa.enceladus.conformance.interpreter.rules.MappingRuleInterpreter=@log.level.testcode@
-log4j.logger.za.co.absa.enceladus.conformance.interpreter.rules.MappingRuleInterpreterGroupExplode=@log.level.testcode@
+log4j.logger.za.co.absa.enceladus.conformance.interpreter.rules.mapping.MappingRuleInterpreter=@log.level.testcode@
+log4j.logger.za.co.absa.enceladus.conformance.interpreter.rules.mapping.MappingRuleInterpreterGroupExplode=@log.level.testcode@
 
 # Suppress a spamming warning from SparkSession$Builder
 log4j.appender.forsparksessionbuilder=org.apache.log4j.ConsoleAppender
