+++ spark-jobs/src/test/scala/za/co/absa/enceladus/conformance/interpreter/rules/RulesSuite.scala
@@ -19,6 +19,7 @@ import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
 import org.apache.spark.sql.types._
 import org.scalatest.FunSuite
+import za.co.absa.enceladus.conformance.interpreter.rules.mapping.{CommonMappingRuleInterpreter, MappingRuleInterpreterGroupExplode}
 import za.co.absa.enceladus.conformance.interpreter.{ExplosionState, InterpreterContextArgs}
 import za.co.absa.enceladus.conformance.samples.EmployeeConformance
 import za.co.absa.enceladus.dao.MenasDAO
@@ -113,48 +114,48 @@ class RulesSuite extends FunSuite with SparkTestBase {
         )
       ))
 
-    MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("id_test", schema, "id", "1")
-    MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("name_test", schema, "name", "'test'")
-    MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("name_null_ok", schema, "name", "null")
-    MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("decimal_test", schema, "price", "1.6127")
-    MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("date_test", schema, "orders.orderdate", "'2017-10-25'")
-    MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("timestamp_test", schema, "orders.delivertime", "'2017-10-25 08:35:43'")
-    MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("boolean_test", schema, "orders.happy", "true")
-    MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("struct_test", schema, "orders.system", "struct('Unknown' as name, 'None' as description)")
+    CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("id_test", schema, "id", "1")
+    CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("name_test", schema, "name", "'test'")
+    CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("name_null_ok", schema, "name", "null")
+    CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("decimal_test", schema, "price", "1.6127")
+    CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("date_test", schema, "orders.orderdate", "'2017-10-25'")
+    CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("timestamp_test", schema, "orders.delivertime", "'2017-10-25 08:35:43'")
+    CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("boolean_test", schema, "orders.happy", "true")
+    CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("struct_test", schema, "orders.system", "struct('Unknown' as name, 'None' as description)")
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("no_attribute_test", schema, "code", "")
+      CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("no_attribute_test", schema, "code", "")
     }.getMessage contains "does not contain the specified target attribute")
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("not_string", schema, "name", "struct('Unknown' as name)")
+      CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("not_string", schema, "name", "struct('Unknown' as name)")
     }.getMessage contains "A string expected")
 
     intercept[ValidationException] {
-      MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("id_null", schema, "id", "null")
+      CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("id_null", schema, "id", "null")
     }
 
     intercept[ValidationException] {
-      MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("id_test", schema, "id", "wrong")
+      CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("id_test", schema, "id", "wrong")
     }
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("decimal_test", schema, "price", "12345.67")
+      CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("decimal_test", schema, "price", "12345.67")
     }.getMessage contains "Scale/precision don't match the value")
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("date_test", schema, "orders.orderdate", "'25/10/2017'")
+      CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("date_test", schema, "orders.orderdate", "'25/10/2017'")
     }.getMessage contains "Make sure the value matches 'yyyy-MM-dd'")
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("timestamp_test", schema, "orders.delivertime", "'25-10-201708:25:43'")
+      CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("timestamp_test", schema, "orders.delivertime", "'25-10-201708:25:43'")
     }.getMessage contains "Make sure the value matches 'yyyy-MM-dd HH:mm:ss'")
 
     intercept[ValidationException] {
-      MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("boolean_test", schema, "orders.happy", "a")
+      CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("boolean_test", schema, "orders.happy", "a")
     }
     intercept[ValidationException] {
-      MappingRuleInterpreterGroupExplode.ensureDefaultValueMatchSchema("struct_test", schema, "orders.system", "struct('unknown' as name)")
+      CommonMappingRuleInterpreter.ensureDefaultValueMatchSchema("struct_test", schema, "orders.system", "struct('unknown' as name)")
     }
 
   }
